Fix a signed/unsigned comparison warning (#27)
The signed value in comparison is always positive, so it's safe to just cast it to unsigned.

+++ guetzli/guetzli.cc
@@ -158,7 +158,7 @@ std::string ReadFileOrDie(FILE* f) {
     exit(1);
   }
   std::unique_ptr<char[]> buf(new char[size]);
-  if (fread(buf.get(), 1, size, f) != size) {
+  if (fread(buf.get(), 1, size, f) != (size_t)size) {
     perror("fread");
     exit(1);
   }
